requirements.txt
addict==2.3.0
anytree==2.6.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
contextlib2==0.6.0.post1
dj-database-url==0.5.0
Django==3.0.8
django-appconf==1.0.4
django-cors-headers==3.4.0
django-decouple==2.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
-e git+https://github.com/pravin-d/drf-autodocs.git@1eb63ac0adac4e94e2d28114f25db9d6a94c2344#egg=drf_autodocs
et-xmlfile==1.0.1
graphviz==0.10.1
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
Markdown==3.3
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
schema==0.7.2
six==1.15.0
sqlparse==0.4.1
treeview==1.2
twilio==6.38.0
urllib3==1.25.10
whitenoise==5.1.0
zipp==3.3.0
